re-organize operator package structure
