+++ api/pkg/controller/operator/master/controller.go
@@ -1,4 +1,4 @@
-package operatormaster
+package master
 
 import (
 	"context"
