+++ api/pkg/controller/operator/master/defaults.go
@@ -1,4 +1,4 @@
-package operatormaster
+package master
 
 import (
 	"fmt"
