+++ api/pkg/controller/operator/master/defaults_test.go
@@ -1,4 +1,4 @@
-package operatormaster
+package master
 
 import (
 	"context"
