+++ api/pkg/controller/operator/master/reconciler.go
@@ -1,4 +1,4 @@
-package operatormaster
+package master
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/kubermatic/kubermatic/api/pkg/controller/operator-master/resources/kubermatic"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/operator/master/resources/kubermatic"
 	operatorv1alpha1 "github.com/kubermatic/kubermatic/api/pkg/crd/operator/v1alpha1"
 	"github.com/kubermatic/kubermatic/api/pkg/resources/reconciling"
