+++ api/pkg/controller/operator/master/resources/kubermatic/util.go
